% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment_01.R
\name{myMean}
\alias{myMean}
\title{myMean()}
\usage{
myMean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
the calculated mean of the vector
}
\description{
myMean()
}
\examples{
x<-c(10,10,10)
myMean(x)
}
